Move metrics aspects to separate starter
