+++ service-1/pom.xml
@@ -121,6 +121,11 @@
             <artifactId>spring-security-test</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>io.github.danielreker.t1-homeworks</groupId>
+            <artifactId>metrics-aspects-spring-boot-starter</artifactId>
+            <version>0.0.12</version>
+        </dependency>
 
     </dependencies>
