--- service-1/src/main/java/io/github/danielreker/t1homeworks/service1/aop/annotation/LogDataSourceError.java
@@ -1,11 +1,0 @@
-package io.github.danielreker.t1homeworks.service1.aop.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface LogDataSourceError {
-}
