--- service-1/src/main/java/io/github/danielreker/t1homeworks/service1/model/dto/DataSourceErrorLogDto.java
@@ -1,17 +1,0 @@
-package io.github.danielreker.t1homeworks.service1.model.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Builder;
-import lombok.Value;
-
-/**
- * DTO for {@link io.github.danielreker.t1homeworks.service1.model.DataSourceErrorLog}
- */
-@Value
-@Builder
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class DataSourceErrorLogDto {
-    String stacktrace;
-    String message;
-    String methodSignature;
-}
