+++ service-1/src/main/java/io/github/danielreker/t1homeworks/service1/service/AccountService.java
@@ -3,9 +3,9 @@ package io.github.danielreker.t1homeworks.service1.service;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.github.danielreker.t1homeworks.service1.aop.annotation.Cached;
-import io.github.danielreker.t1homeworks.service1.aop.annotation.Metric;
+import io.github.danielreker.t1homeworks.metricsaspectsspringbootstarter.aop.annotation.Metric;
 import io.github.danielreker.t1homeworks.service1.model.dto.AccountDto;
-import io.github.danielreker.t1homeworks.service1.aop.annotation.LogDataSourceError;
+import io.github.danielreker.t1homeworks.metricsaspectsspringbootstarter.aop.annotation.LogDataSourceError;
 import io.github.danielreker.t1homeworks.service1.mapper.AccountMapper;
 import io.github.danielreker.t1homeworks.service1.model.Account;
 import io.github.danielreker.t1homeworks.service1.repository.AccountRepository;
