+++ service-1/src/main/java/io/github/danielreker/t1homeworks/service1/service/ClientService.java
@@ -3,8 +3,8 @@ package io.github.danielreker.t1homeworks.service1.service;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.github.danielreker.t1homeworks.service1.aop.annotation.Cached;
-import io.github.danielreker.t1homeworks.service1.aop.annotation.LogDataSourceError;
-import io.github.danielreker.t1homeworks.service1.aop.annotation.Metric;
+import io.github.danielreker.t1homeworks.metricsaspectsspringbootstarter.aop.annotation.LogDataSourceError;
+import io.github.danielreker.t1homeworks.metricsaspectsspringbootstarter.aop.annotation.Metric;
 import io.github.danielreker.t1homeworks.service1.mapper.ClientMapper;
 import io.github.danielreker.t1homeworks.service1.model.Client;
 import io.github.danielreker.t1homeworks.service1.model.dto.ClientDto;
